feat: 커스텀 타이머 실행 로직 구현 #227

+++ domain/src/main/java/com/project200/domain/model/TimerModel.kt
@@ -14,6 +14,6 @@ data class SimpleTimer(
 data class Step(
     val id: Long,
     val order: Int,
-    val time: Int,
+    val time: Long,
     val name: String
 )
